+++ ocdskit/cli/commands/package_records.py
@@ -1,5 +1,6 @@
 from ocdskit.cli.commands.base import OCDSCommand
 from ocdskit.combine import package_records
+from ocdskit.util import grouper
 
 
 class Command(OCDSCommand):
@@ -8,6 +9,7 @@ class Command(OCDSCommand):
 
     def add_arguments(self):
         self.add_argument('extension', help='add this extension to the package', nargs='*')
+        self.add_argument('--size', type=int, help='the maximum number of records per package')
 
         self.add_package_arguments('record')
 
@@ -15,6 +17,10 @@ class Command(OCDSCommand):
         kwargs = self.parse_package_arguments()
         kwargs['extensions'] = self.args.extension
 
-        output = package_records(self.items(), **kwargs)
-
-        self.print(output, streaming=True)
+        if self.args.size:
+            for data in grouper(self.items(), self.args.size):
+                output = package_records(filter(None, data), **kwargs)
+                self.print(output, streaming=True)
+        else:
+            output = package_records(self.items(), **kwargs)
+            self.print(output, streaming=True)